Bunch Upload

+++ sketches/logik/editor.js
@@ -18,7 +18,6 @@ class Editor {
   constructor() {
     this.mouse;
     this.outputSocket;
-    this.hoverSocket;
     this.dragLogic;
   }
 
@@ -34,7 +33,7 @@ class Editor {
     for (var i = 0; i < logics.length; i++) {
       var logic = logics[i];
       //this shoudl really use rect detection
-      if (logic.pos.dist(this.mouse) < 10) {
+      if (logic.pos.dist(this.mouse) < 20) {
         this.dragLogic = logic;
       }
     }
@@ -49,7 +48,7 @@ class Editor {
     for (var i = 0; i < sockets.length; i++) {
       var socket = sockets[i];
 
-      if (socket.pos.dist(this.mouse) < socket.width / 2) {
+      if (socket.pos.dist(this.mouse) < socket.width) {
         // Select from socket or create connection
 
         // select output
@@ -58,7 +57,8 @@ class Editor {
         }
 
         //delete connection, create new
-        else if (this.outputSocket == null && socket instanceof InputSocket) {
+        else if (this.outputSocket == null && socket instanceof InputSocket
+        && socket.connections[0] != null) {
           this.outputSocket = socket.connections[0].input;
           socket.connections[0].delete();
         }
@@ -69,6 +69,7 @@ class Editor {
           var c1 = new Connection(this.outputSocket, socket);
           this.outputSocket = null;
         }
+
         break;
       }
     }
@@ -76,15 +77,6 @@ class Editor {
 
   update() {
     this.mouse = createVector(mouseX, mouseY);
-
-    this.hoverSocket = null;
-    for (var i = 0; i < sockets.length; i++) {
-      var socket = sockets[i];
-      if (socket.pos.dist(this.mouse) < socket.width) {
-        this.hoverSocket = socket;
-        break;
-      }
-    }
 
     if (this.dragLogic != null) {
       this.dragLogic.pos = this.mouse;
@@ -95,22 +87,37 @@ class Editor {
 
   draw() {
 
-    if(this.hoverSocket != null) {
-      strokeWeight(2);
-
-      if((this.outputSocket == null && this.hoverSocket instanceof InputSocket)
-      || (this.outputSocket != null && this.hoverSocket instanceof OutputSocket)
-      || !this.hoverSocket.canEstablishConnection()) {
-        stroke(200, 0, 0);
-      } else {
-        stroke(0, 200, 0);
+    // Check if hovering socket and draw
+    for (var i = 0; i < sockets.length; i++) {
+      var socket = sockets[i];
+      if (socket.pos.dist(this.mouse) < socket.width) {
+        strokeWeight(2);
+        noFill();
+        if((this.outputSocket == null && socket instanceof InputSocket)
+        || (this.outputSocket != null && socket instanceof OutputSocket)
+        || !socket.canEstablishConnection()) {
+          stroke(200, 0, 0);
+        } else {
+          stroke(0, 200, 0);
+        }
+        rect(socket.pos.x, socket.pos.y, socket.width, socket.height);
+        break;
       }
-
-      noFill();
-      rect(this.hoverSocket.pos.x, this.hoverSocket.pos.y,
-      this.hoverSocket.width, this.hoverSocket.height);
     }
 
+    // Check if hovering logic and draw
+    for (var i = 0; i < logics.length; i++) {
+      var logic = logics[i];
+      if (logic.pos.dist(this.mouse) < 20) {
+        stroke(color(112, 2, 124));
+        strokeWeight(logic.frameWidth);
+        noFill();
+        rect(logic.pos.x, logic.pos.y, logic.width, logic.height);
+        break;
+      }
+    }
+
+    // Temporary Connection Draw
     if (this.outputSocket != null) {
       strokeWeight(6);
       stroke(0, 100, 0);
